requirements.txt
openai
python-dotenv
requests
requests-oauthlib
pytz
python-telegram-bot